requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
mysql-connector-python==8.1.0
python-multipart==0.0.6
jinja2==3.1.2
SpeechRecognition==3.10.0
python-dateutil==2.8.2
pydantic==2.4.2
websockets==12.0
groq
phidata